requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-cors-headers==2.4.0
django-environ==0.4.5
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
ed25519==1.4
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.3.13
openapi-codec==1.3.2
Pygments==2.2.0
python-dateutil==2.7.5
pytz==2018.5
requests==2.21.0
simplejson==3.16.0
six==1.12.0
Token==1.8.6
uritemplate==3.0.0
urllib3==1.24.1
